scripts/TSDB/backend/requirements.txt
fastapi
influxdb
uvicorn
python-dotenv